[v0.0.6.11] - removed additional x, y, size components - changed hierarchy of GUI
