+++ src/main/java/net/archasmiel/thaumcraft/screen/thaumonomicon/ThaumonomiconGui.java
@@ -2,9 +2,9 @@ package net.archasmiel.thaumcraft.screen.thaumonomicon;
 
 import io.github.cottonmc.cotton.gui.client.LightweightGuiDescription;
 import net.archasmiel.thaumcraft.screen.thaumonomicon.data.Tabs;
-import net.archasmiel.thaumcraft.screen.thaumonomicon.parts.Gui;
-import net.archasmiel.thaumcraft.screen.thaumonomicon.parts.Panel;
-import net.archasmiel.thaumcraft.screen.thaumonomicon.parts.Tab;
+import net.archasmiel.thaumcraft.screen.thaumonomicon.parts.researchview.ResearchGUI;
+import net.archasmiel.thaumcraft.screen.thaumonomicon.parts.researchview.Panel;
+import net.archasmiel.thaumcraft.screen.thaumonomicon.parts.researchview.Tab;
 
 import java.util.List;
 
@@ -12,25 +12,17 @@ import static net.archasmiel.thaumcraft.screen.thaumonomicon.data.Tabs.*;
 
 public class ThaumonomiconGui extends LightweightGuiDescription {
 
+    public ThaumonomiconGui() {
+        setRootPanel(root);
+    }
     private static final Integer SIZE_X = 256 + TAB_SIZE;
     private static final Integer SIZE_Y = 256 - TAB_SIZE;
 
     private static final Integer RESEARCH_X = SIZE_X - TAB_SIZE;
     private static final Integer RESEARCH_Y = SIZE_Y;
 
-
     private final Panel researchPanel = new Panel(TAB_SIZE, 0, RESEARCH_X, RESEARCH_Y);
-    private final Gui root = new Gui(researchPanel, Tabs.getTabList(), SIZE_X, SIZE_Y);
-
-
-
-
-
-
-    public ThaumonomiconGui() {
-        setRootPanel(root);
-    }
-
+    private final ResearchGUI root = new ResearchGUI(researchPanel, Tabs.getTabList(), SIZE_X, SIZE_Y);
 
     public List<Tab> getTabs() {
         return root.getTabs();
