+++ src/main/java/net/archasmiel/thaumcraft/screen/thaumonomicon/parts/bookview/BookGUI.java
@@ -1,0 +1,9 @@
+package net.archasmiel.thaumcraft.screen.thaumonomicon.parts.bookview;
+
+import io.github.cottonmc.cotton.gui.widget.WPlainPanel;
+
+public class BookGUI extends WPlainPanel {
+
+
+
+}
